Adds exception handling and image on Readme

+++ src/main/java/command/DeleteCommand.java
@@ -33,6 +33,10 @@ public class DeleteCommand extends Command {
      */
     public String execute(TaskList tasks, Ui ui, Storage storage) {
         try {
+            if (this.taskNumber > tasks.size()) {
+                return Ui.showException(new DukeException(
+                        "☹ OOPS!!! You only have " + tasks.size() + " tasks in the list, you cannot delete task number " + taskNumber));
+            }
             Task removedTask = tasks.get(taskNumber - 1);
             tasks.remove(removedTask);
             storage.deleteFromFile(taskNumber);
